views/browse: split into outline-render + show-landing + pager-crawl submodules — #98
admin/app/views/browse.js was 2270 lines doing six things. Three peel
into peer submodules under admin/app/views/browse/:

  - outline-render.js — entry → DOM (renderEntry, renderOutline,
    renderSection, renderFlatSection, renderCard, drillHashFor,
    drillPartsFor, drillPartsForUrl, primeTuneinSkipCaches, related-
    chip + pivot-chip renderers, localCountry lift).
  - show-landing.js — c=pbrowse Describe + Browse composite
    (loadShowLanding, renderShowLandingCard, renderLiveShowCard,
    renderTopicsCard, renderTopicRow, plus _renderShowLandingForTest).
  - pager-crawl.js — per-section Load-more pagers + the eager-serial
    filter-crawl coordinator + the strap. Owns _activePagers,
    appendNewRows, disposeActivePagers, the filter input handle,
    applyDomFilter, currentFilterQuery, ensureCoordinator.

views/browse.js shrinks to 760 lines covering only root tabs, drill
orchestration (renderDrill / renderShowLanding share a buildDrillFrame
helper), crumb stack + hydration, filter input mount, and lifecycle
(disposeActivePagers from the submodule). Every existing public
surface — renderEntry / renderOutline / mountLoadMoreButtons /
filterRowEntries / parseCrumbs / stringifyCrumbs / crumbTokenFor /
partsFromCrumb / crumbLabelFor / backHrefFor / renderCrumbTrail and
the test hooks — re-exports from views/browse.js so existing imports
keep working.

The cache.set(tuneinTopicNameKey(gid), text, TTL_LABEL) write in
primeTuneinSkipCaches (#102) is preserved, as is the c=pbrowse
handling and topic-row pattern (#84/#87).

Test split mirrors the module split: test_browse.js keeps the top-
level coverage; new test_browse_outline.js / test_browse_show_landing.js /
test_browse_pager_crawl.js exercise each submodule directly through
its own imports. 672/672 → 723/723.
